# openapi.json
{"openapi":"3.1.0","info":{"title":"Sup de Vinci Travel Hub","summary":"Plate-forme B2C qui aggrège vols, hébergements et activités touristiques afin de construire des itinéraires personnalisés quasi en temps réel.","version":"0.1.0"},"paths":{"/login":{"post":{"tags":["auth"],"summary":"Creates a session","description":"Creates a session from an arbitrary user_id passed in the body.","operationId":"login_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginInfo"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cities/":{"get":{"tags":["city"],"summary":"Get a paginated list of cities.","operationId":"get_offers_cities__get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","exclusiveMinimum":0,"default":1,"title":"Page"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","exclusiveMinimum":0,"default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["city"],"summary":"Create a city","operationId":"new_city_cities__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/City"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cities/{city_code}":{"get":{"tags":["city"],"summary":"Get a single city by its code.","operationId":"get_single_city_cities__city_code__get","parameters":[{"name":"city_code","in":"path","required":true,"schema":{"type":"string","title":"City Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseModel_City_"}}}},"404":{"description":"City not found","content":{"application/json":{"example":{"detail":"City not found"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/offers/":{"get":{"tags":["offers"],"summary":"Get a paginated list of offers.","operationId":"get_offers_offers__get","parameters":[{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string","minLength":3},{"type":"null"}],"title":"From"}},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string","minLength":3},{"type":"null"}],"title":"To"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","exclusiveMinimum":0,"default":1,"title":"Page"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","exclusiveMinimum":0,"default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["offers"],"summary":"Create an offer","operationId":"new_offer_offers__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/app__entities__offer__Offer-Input__1"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/offers/{id}":{"get":{"tags":["offers"],"summary":"Get a single offer by its ID.","operationId":"get_single_offer_offers__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"City not found","content":{"application/json":{"example":{"detail":"City not found"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reco":{"get":{"tags":["recommendations"],"summary":"Get recommandations for a city","operationId":"get_recommandations_reco_get","parameters":[{"name":"city","in":"query","required":true,"schema":{"type":"string","description":"The 3-letter code of the city","title":"City"},"description":"The 3-letter code of the city"},{"name":"k","in":"query","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"description":"The number of recommendations to return","title":"K"},"description":"The number of recommendations to return"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"City":{"properties":{"code":{"type":"string","minLength":3,"title":"Code"},"name":{"type":"string","title":"Name"},"country":{"type":"string","maxLength":3,"minLength":2,"title":"Country"}},"type":"object","required":["code","name","country"],"title":"City"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginInfo":{"properties":{"userId":{"type":"string","title":"Userid"}},"type":"object","required":["userId"],"title":"LoginInfo"},"ResponseModel_City_":{"properties":{"data":{"$ref":"#/components/schemas/City"}},"type":"object","required":["data"],"title":"ResponseModel[City]"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}